DRY out labels function
