+++ src/smallgroups/CSGroups.js
@@ -1,5 +1,6 @@
 import Base from '../base'
 import Group from './group'
+import { buildLabels } from '../components/labels'
 
 export default class CSGroups extends Base {
 	buildModelObject = function (model) {
@@ -11,38 +12,7 @@ export default class CSGroups extends Base {
 		this.sites.sort()
 
 		// loop the labels and capture them
-		if (this.options.hasOwnProperty('show_labels')) {
-			let labelIds = Object.keys(this.label)
-			let labelsIds = this.labels.map(l => '' + l.id)
-
-			this.options.show_labels.forEach(labelId => {
-				// try and find this label on the model
-				model.labels.forEach(label => {
-					if (label.id == labelId) {
-						// don't add a label object unless we need it by id
-						if (!labelsIds.includes('' + label.id)) {
-							this.labels.push({
-								id: label.id,
-								multiple: label.multiple,
-								name: label.name,
-								options: label.options,
-								required: label.required,
-							})
-						}
-						// don't add a label filter value unless we need it by id
-						if (!labelIds.includes('' + label.id)) {
-							this.label[labelId] = {
-								id: labelId,
-								value: null,
-							}
-							this.$watch('label[' + labelId + '].value', v => {
-								this.filterModels()
-							})
-						}
-					}
-				})
-			})
-		}
+		this.buildLabels()
 
 		return new Group(model)
 	}
@@ -237,5 +207,8 @@ export default class CSGroups extends Base {
 		this.tag = null // tag string for filterModels()
 		this.tagOptions = []
 		this.tags = [] // @deprecated tags names array
+
+		// shared function between label-using classes
+		this.buildLabels = buildLabels
 	}
 }
